Interpreter Merge
I think the new interpreter system has been successfully integrated to work with all current blocks (not including IMU, Distance Sensor, Acceleration, and Orientation Blocks. These blocks are still on the do not use list.) I did some tests and ran into no bugs so far, but it has become a big system to test with this being a small rebuild of the fundamental system.

MotorProperty, Gamepad, and Telemetry System Changed to work with new system.

The interpreter seems to run much faster with decent delay still in effect when you have the console open. I don't know if that is an issue we should be warry of or not.

I also changed the linearOpMode color as I'm colorblind and it was a close color to the sensors category.